+++ CHANGELOG.md
@@ -8,45 +8,101 @@ For each release, the changelog describes what was added, changed, fixed, or rem
 
 At the end of this file, the release history is provided with links to previous releases.
 
-## 1.6.0 - 2026-08-27
+## 1.7.0 - 2026-08-27
 
-Class & Attribute Update
+FileSystem Update
 
 ### Runtime
 
-- Improved property evaluation.
-- Added string properties:
-    - length
-    - as_upper
-    - as_lower
-- Added internal length getter for String.
-- Added playscript().
+- Added FileSystem.
+- Added FileSystem.File.
+- Added FileSystem.Folder.
+- Added FileSystem.tempfolder.
+- Added FileSystem.File.exists().
+- Added FileSystem.File.read().
+- Added FileSystem.File.write().
+- Added FileSystem.File.store().
+- Added FileSystem.File.append().
+- Added FileSystem.Folder.exists().
+- Added FileSystem.Folder.open().
+- Added FileSystem.Folder.add_file().
+- Added FileSystem.append_to_file().
+- Added FileSystem.write_to_file().
+- Added FileSystem.exists().
+- Added FileSystem.folder_exists().
+- Added FileSystem.file_exists().
+- Added FileSystem.open_folder().
+- Added FileSystem.add_file_to_folder().
+- Added FileSystem.read_file().
+- Added printfinish().
+- Added print(printtype = "successful") support.
+- Added F6 shortcut for running GameScript files.
 
 ### Editor
 
-- Added dynamic semantic highlighting for variables.
-- Added dynamic semantic highlighting for properties.
-- Added highlighting for:
-    - .length
-    - .as_upper
-    - .as_lower
-- Improved property highlighting without requiring every property to be hard-coded in the TextMate grammar.
-
-### Parser
-
-- Improved nested expression and property parsing.
-- Added support for finding attribute dots outside parentheses and strings.
-- Added support for finding matching parentheses when evaluating expressions.
+- Added highlight for printfinish().
+- Added highlight for FileSystem.
+- Added highlight for FileSystem.File.
+- Added highlight for FileSystem.Folder.
+- Added highlight for FileSystem.tempfolder.
+- Added highlight for FileSystem.File.exists().
+- Added highlight for FileSystem.File.read().
+- Added highlight for FileSystem.File.write().
+- Added highlight for FileSystem.File.store().
+- Added highlight for FileSystem.File.append().
+- Added highlight for FileSystem.Folder.exists().
+- Added highlight for FileSystem.Folder.open().
+- Added highlight for FileSystem.Folder.add_file().
+- Added highlight for FileSystem.append_to_file().
+- Added highlight for FileSystem.write_to_file().
+- Added highlight for FileSystem.exists().
+- Added highlight for FileSystem.folder_exists().
+- Added highlight for FileSystem.file_exists().
+- Added highlight for FileSystem.open_folder().
+- Added highlight for FileSystem.add_file_to_folder().
+- Added highlight for FileSystem.read_file().
+- Added printfinish() to CompletionItems.
+- Added FileSystem to CompletionItems.
+- Added FileSystem.File to CompletionItems.
+- Added FileSystem.Folder to CompletionItems.
+- Added FileSystem.tempfolder to CompletionItems.
+- Added FileSystem.File.exists() to CompletionItems.
+- Added FileSystem.File.read() to CompletionItems.
+- Added FileSystem.File.write() to CompletionItems.
+- Added FileSystem.File.store() to CompletionItems.
+- Added FileSystem.File.append() to CompletionItems.
+- Added FileSystem.Folder.exists() to CompletionItems.
+- Added FileSystem.Folder.open() to CompletionItems.
+- Added FileSystem.Folder.add_file() to CompletionItems.
+- Added FileSystem.append_to_file() to CompletionItems.
+- Added FileSystem.write_to_file() to CompletionItems.
+- Added FileSystem.exists() to CompletionItems.
+- Added FileSystem.folder_exists() to CompletionItems.
+- Added FileSystem.file_exists() to CompletionItems.
+- Added FileSystem.open_folder() to CompletionItems.
+- Added FileSystem.add_file_to_folder() to CompletionItems.
+- Added FileSystem.read_file() to CompletionItems.
+- Added printfinish() hover support.
+- Added FileSystem hover support.
+- Added FileSystem.File hover support.
+- Added FileSystem.Folder hover support.
+- Added FileSystem.tempfolder hover support.
+- Added FileSystem.File.exists() hover support.
+- Added FileSystem.File.read() hover support.
+- Added FileSystem.File.write() hover support.
+- Added FileSystem.File.store() hover support.
+- Added FileSystem.File.append() hover support.
+- Added FileSystem.Folder.exists() hover support.
+- Added FileSystem.Folder.open() hover support.
+- Added FileSystem.Folder.add_file() hover support.
 
 ### Demos
 
-- Added demos folder.
-- Added StringData.gs demo.
-- Added StringRepeater.gs demo.
-- Added NumberTest.gs demo.
+- Added FileHelper.gs demo.
 
 ## Releases
 
+- [1.7.0](https://github.com/qzou1222-alt/GameScript/blob/v1.7.0/CHANGELOG.md)
 - [1.6.0](https://github.com/qzou1222-alt/GameScript/blob/v1.6.0/CHANGELOG.md)
 - [1.5.0](https://github.com/qzou1222-alt/GameScript/blob/v1.5.0/CHANGELOG.md)
 - [1.4.0](https://github.com/qzou1222-alt/GameScript/blob/v1.4.0/CHANGELOG.md)